move javadoc and minor cleanup
