+++ apm-agent-core/src/main/java/co/elastic/apm/agent/objectpool/Recyclable.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -26,5 +26,8 @@ package co.elastic.apm.agent.objectpool;
 
 public interface Recyclable {
 
+    /**
+     * resets pooled object state so it can be reused
+     */
     void resetState();
 }
