+++ apm-agent-core/src/main/java/co/elastic/apm/agent/objectpool/Resetter.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -22,14 +22,27 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.objectpool.impl;
+package co.elastic.apm.agent.objectpool;
 
-import co.elastic.apm.agent.objectpool.Recyclable;
-
+/**
+ * Defines resetter that is responsible to reset pooled object state
+ *
+ * @param <T> pooled object type
+ */
 public interface Resetter<T> {
 
+    /**
+     * Recycles a pooled object state
+     *
+     * @param object object to recycle
+     */
     void recycle(T object);
 
+    /**
+     * Resetter for objects that implement {@link Recyclable}
+     *
+     * @param <T> recyclable object type
+     */
     class ForRecyclable<T extends Recyclable> implements Resetter<T> {
         private static ForRecyclable INSTANCE = new ForRecyclable();
